requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.0
huggingface-hub==0.31.4
idna==3.10
kagglehub==0.3.12
numpy==2.2.6
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.2
typing_extensions==4.13.2
urllib3==2.4.0
